Fix commit tab crash on repos with long histories
Specifically, repos with more than a month from first to last commit
would crash because of a misunderstanding of what 'date' is. The code
checked that the date did not exceed the end date, but the 'date' is
merely the day of the month (1 to 31), not the entire
year-month-qualified date, so the code didn't process commits with a
later date in the month than the end date. Fixed by merely removing that
safeguard-check.

Closes #7